fix: get /api/oauth/id-token is implemented inline
